shared-library/pyproject.toml
[project]
name = "stockdice"
version = "0.1.0"
description = "A random stock picker working with FMP API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.40.3",
    "httpx>=0.28.1",
    "pandas>=2.3.1",
    "requests>=2.32.4",
    "toml>=0.10.2",
]
